myaccount/models.py
# Create your models here.
from allauth.account.models import EmailAddress
from django.contrib.auth.models import User
from django.db import models

from wechat.models import WechatUserInfo


class UserProfile(models.Model):
    """

    """
    user = models.OneToOneField(User, on_delete=models.CASCADE, related_name='profile')

    org = models.CharField(
        'Organization', max_length=128, blank=True)

    telephone = models.CharField(
        'Telephone', max_length=50, blank=True)

    mod_date = models.DateTimeField('Last modified', auto_now=True)

    wechat = models.ForeignKey(WechatUserInfo, on_delete=models.DO_NOTHING, null=True, blank=True)

    class Meta:
        verbose_name = 'User Profile'

    def __str__(self):
        return "{}'s profile".format(self.user.__str__())

    def account_verified(self):
        if self.user.is_authenticated:
            result = EmailAddress.objects.filter(email=self.user.email)
            if len(result):
                return result[0].verified
        return False
