requirements.txt
django
djangorestframework
django-rest-swagger
django_filters
django-allauth
python-weixin
pypinyin
djangorestframework-queryfields